requirements.txt
Django==5.0.2
psycopg2==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-phonenumber-field[phonenumbers]
GDAL
channels
daphne
channels_redis
django-enumchoicefield
geoalchemy2
pandas
